+++ src/app/screens/Home/index.tsx
@@ -9,8 +9,8 @@ import Header from "@components/Header";
 import IconButton from "@components/IconButton";
 import Loading from "@components/Loading";
 import Progressbar from "@components/Progressbar";
+import PublisherCard from "@components/PublisherCard";
 import TransactionsTable from "@components/TransactionsTable";
-import PublisherCard from "@components/new/PublisherCard";
 import { Tab } from "@headlessui/react";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
