+++ src/app/screens/LNURLPay.tsx
@@ -1,10 +1,10 @@
 import Button from "@components/Button";
+import ConfirmOrCancel from "@components/ConfirmOrCancel";
 import Container from "@components/Container";
+import PublisherCard from "@components/PublisherCard";
 import SatButtons from "@components/SatButtons";
 import DualCurrencyField from "@components/form/DualCurrencyField";
 import TextField from "@components/form/TextField";
-import ConfirmOrCancel from "@components/new/ConfirmOrCancel";
-import PublisherCard from "@components/new/PublisherCard";
 import axios from "axios";
 import React, { useState, useEffect, MouseEvent } from "react";
 import { useNavigate, useSearchParams } from "react-router-dom";
